anime-rec-engine/ml-service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
scikit-learn==1.7.1
pandas==2.3.1
numpy==2.3.1
sentence-transformers==5.0.0
openai==1.97.0
python-dotenv==1.1.1
pydantic==2.11.7
httpx==0.28.1
redis==5.0.1
psycopg2-binary==2.9.9
elasticsearch==8.11.0
pytest==7.4.3
pytest-asyncio==0.21.1 